Minor refactoring to prevent deferred imports
